cleanup: autoformats and minor visual fixes
